backend/requirements.txt
# Backend API server
fastapi==0.115.6
uvicorn==0.34.0
python-dotenv==1.0.1

# ML dependencies (used by src/ modules called via run_tool.py)
pandas==2.2.3
numpy==2.2.3
mlflow==2.21.3
scikit-learn==1.6.1
